requirements.txt
Scrapy==2.12.0
scrapy-playwright==0.0.42
playwright==1.49.1
python-dotenv==1.0.1
mysql-connector-python==9.1.0
mysqlclient==2.2.7